documentos

+++ db/db_file.py
@@ -2,6 +2,7 @@ from schemas import FileBaseModel
 from sqlalchemy.orm.session import Session
 from db.models import DbFile
 from fastapi import HTTPException
+from fastapi import status
 
 from abc import ABC, abstractmethod
 
@@ -70,10 +71,15 @@ class CrudFile(CrudFileInterface):
         return file
 
     @staticmethod
-    def delete_file(id, db: Session):
+    def delete_file(id:int, db: Session, user_id: int):
         file = db.query(DbFile).filter(DbFile.id == id).first()
+        
         if not file:
-            raise HTTPException(status_code=404, detail="File not found")
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="File not found")
+        
+        if file.user_id != user_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You can't delete this file,only the owner can")
+        
         db.delete(file)
         db.commit()
-        return file
+        return "ok"
